main aufgeräumt und erode implementiert
